DP_Federated/requirements.txt
absl-py==1.0.0
astunparse==1.6.3
attrs==21.4.0
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
cloudpickle==2.2.1
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
dill==0.3.6
dm-tree==0.1.8
dp-accounting==0.4.1
farmhashpy==0.4.0
flatbuffers==1.12
fonttools==4.39.4
gast==0.4.0
google-auth==2.19.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.59.0
grpcio==1.54.2
h5py==3.8.0
idna==3.4
importlib-metadata==6.6.0
jax==0.3.24
jaxlib==0.3.24
joblib==1.2.0
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.6.2
mpmath==1.2.1
numpy==1.22.4
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
pandas==1.5.1
Pillow==9.5.0
portpicker==1.5.0
promise==2.3
protobuf==3.19.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytorch-tabnet==4.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.1.3
scipy==1.7.3
semantic-version==2.6.0
six==1.16.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-datasets==4.5.2
tensorflow-estimator==2.9.0
tensorflow-federated==0.34.0
tensorflow-macos==2.9.2
tensorflow-metadata==1.13.1
tensorflow-metal==0.5.1
tensorflow-model-optimization==0.7.3
tensorflow-privacy==0.8.4
tensorflow-probability==0.15.0
termcolor==2.3.0
threadpoolctl==3.1.0
torch==1.13.1
tqdm==4.65.0
typing_extensions==4.6.3
urllib3==1.26.16
Werkzeug==2.3.4
wget==3.2
wrapt==1.15.0
zipp==3.15.0
